feat: add `flox-activations executive`
Rather than calling start.bash and starting services from
`flox-activations activate`, spawn a child `flox-activations executive`
that sends SIGUSR1 when start completes. `flox-activations activate`
waits for SIGUSR1 and then proceeds to attach when SIGUSR1 is received.

For now, `flox-activations executive` handles logging exactly the same
as `flox-activations activate`; we'll probably want to redirect logs to
a file as followup.

For now, `flox-activations activate` starts services when attaching,
which could start a new instance of process-compose. We'll want to
instead communicate with an existing process-compose as followup.