feat: substitute custom packages in parallel
Custom packages are now downloaded in parallel. Making this work
properly with authentication turned out to be a pain, so instead we
eagerly try to create a `netrc` file, but don't inspect whether that
succeeded or not until requested. If it was requested and didn't
succeed, then we error at the time that it's inspected.