feat: download all packages in parallel
This change implements full parallelism when downloading packages rather
than using parallelism only within each kind of package.

During a demo it was mentioned that users with poor internet may hit
timeouts in Nix if we're trying to download all packages at the same
time. To begin to address that I added a semaphore with the intention
that a follow-up commit will provide the configuration mechanism for
the number of reservations allowed by the semaphore. Currently we still
spawn all threads at the same time and pass the semaphore into each
thread so that the thread can release the semaphore when the download
is complete. There's not currently an easy way to do this in such a way
that threads aren't spawned until they're needed.