feat: add variable for parallel downloads
This change adds a new variable `FLOX_MAX_PARALLEL_DOWNLOADS` that a
user can use to specify the maximum number of active downloads they'd
like. This can be useful in cases where they're on a poor connection.
In that scenario, many parallel downloads may saturate their connection,
cause individual downloads to take longer, and potentially cause
timeouts in Nix as it does the underlying substitution (it would be
good to confirm that these timeouts are still possible).

I debated about where to put this. A user on a poor connection could
just be at a coffee shop for the moment, in which case persisting this
value to config feels like it limits the times where they aren't at a
coffee shop. I'm also not in love with putting it on the `Flox` struct
since that requires developers to remember to pass this value down to `
build` at all of the various places that it's called. There's a wealth
of places that cause an environment rebuild, so it actually feels
cleaner to read this environment variable in the one places that it's
used.