feat(executive): Call watchdog as library
Re-uses the existing watchdog code as a library to start the monitoring
loop and GC log threads.

We switch to file logging and close file descriptors immediately before,
so that it doesn't spam the current activation terminal, especially if
you've specified `-v` levels, and so that we have a record of logs when
the long-lived executive is running in the background.

Removes all of the machinery to build a separate `flox-watchdog` binary.