chore(watchdog): Make log dir and socket mandatory
It doesn't look like we've ever invoked the watchdog without the
`--log-dir`, even for tests.

The socket is always passed, even when services weren't present in the
manifest or started, because they could be added at a later date and the
watchdog needed to attempt to shutdown process-compose on exit.