feat: retrieve generations metadata at the (cached) remote state
The remote state of environments is stored on a separate "sync branch".
We dont want to directly expose the sync branch itself
as its curently only used to decouple (hashed) instance branches
of environments from the remote synchronization.
This allows checking properties of the remote using local git operations.
In this case we allow acess to the remote (i.e. FloxHub)
generation metadata.
The metadata will be used to compare against
the respective local state to notify of state differences
during activation.

NOTE: currently this requires the remote state to be explicitly fetched
before calling this function to retrieve "fresh" remote state.
We could fetch here too which could leak environment types (errors) into
the so far isolated `Generation` module.
I opted for the semaration to keep the interface overlap low.