feat(upgrade-checks): fetch last known generations state on FloxHub
* add optional generations metadata of the remote state (i.e. on FloxHub)
When checking managed environments (`Managed-` or `RemoteEnvironment`),
also attempt to fetch their respective state on FloxHub.
This state can then be comapared with the local state
to notify users of pending changes on either side.

* Rename `UpgradeInformation::result` field to `upgrade_result`
`UpgradeInformation` used to only cover "package upgrades",
in which case `result` was referring to `UpgradeResult`s
and was somewhat less ambiguous.
With the addition of generation metadata `result` becomes less clear.