test(services): Remove watchdog.bats + move 1 test
Most of this functionality is hard to test now that the watchdog has
been inlined into `flox-activations executive` and it doesn't seem worth
the effort when we have reasonable unit test coverage for signal and PID
handling.

The one exception is ensuring that imperatively started services are
shutdown when activations exit, which I've moved to `services.bats` with
a minor modification to look more like the other shutdown tests and not
duplicate the function from `start_shuts_down_process_compose.sh`.