test(bats): wait_for_{watchdogs -> activations}
In the move from `flox-watchdog` to `flox-activations executive` we no
longer have the project directory available to match in the process
arguments.

The "north star" branch of the activation refactor used grep/tail/sed to
parse the PID out but I'm concerned that might be brittle.

A more reliable approach is to observe `activations.json` which gets
added to by `flox-activations activate` early in the activation process
and cleaned up by `flox-activations executive` at the very end. This
also solves our long running problem of false-negatives when the check
runs before the watchdog has started.