test(watchdog): Cleanup activations.json changes
I'm improving `wait_for_watchdogs` in a subsequent commit to check the
contents of `activations.json` rather than the watchdog process.

This revealed that the watchdog sometimes exits early in these tests
because it detects the version mismatch in the polling loop and doesn't
perform any cleanup.

Recover from this from happening by resetting our changes before moving
on to the next test. No services are started so we don't need to clean
them up.