fix(containerize): Provide fake log+socket paths
`flox containerize` was broken by making the log and socket paths of
`ActivateCtx` no longer optional because it's the only use that doesn't
have them. These aren't used when `run_monitoring_loop: false`

Rather than make them optional again, which require wrapping and
unwrapping them in all of the other cases where they are present, I
think it's clearer for now to provide "fake" values.

Later on it would be preferable to combine these options in a single
variant so that it's clearer how they depend on each other.