feat(executive): Enable sub-reaper on Linux
Michael has requested that any orphaned grandchild process of
on-activate hooks and services are re-parented and reaped by the
executive, rather than being automatically re-parented and reaped by PID
1, so that they stay within our process tree and the user can tell that
they originated from us.

We can only do this on Linux because there's no equivalent to the
sub-reaper on Darwin. This is considered OK because it's more of a
concern for production use which is more likely to happen on Linux.

The guard will ensure that we reap any child process on exit,
particularly if we exit early due to an error. We could have additional
reap calls between activate hooks and services but there's no real
urgency to cleaning up zombies early.

I will add a SIGCHLD handler in a subsequent commit that will reap
processes on-demand once we enter the monitoring loop.