feat(executive): Monitoring loop reap on SIGCHLD
This compliments the sub-reaper guard in the executive by reaping any
children when the kernel tells us they have exited and been re-parented
to us.

I've put it at the end of the monitoring loop match statement because
there's less urgency to clean up zombies than there is in responding to
activations exiting.

The call is effectively a noop on Darwin because we're not creating any
child processes to wait on.

Per the warning in the code comment, this could cause us problems if we
use `Command::wait` (or `Command::output` or `Command::status`)
concurrently to the monitoring loop because the processes could be
reaped before we wait, as we've seen in the Kubernetes shim.

We've not currently affected though because PID checking happens in a
separate part of the loop and GCing logs only uses file operations.