feat: format errors in flox-activations
Format errors in flox-activations with the `❌ ERROR` prefix.

Since flox-activations now formats errors, the CLI has to defer to it
completely for printing errors for ephemeral activations, which is an
improvement because it avoids printing:
```
❌ ERROR: ... Error: ...
```

I think this formatting is clearly better for `activate` and `executive`
subcommands as well as all cases covered by our tests. I'm not
completely sure if there are any places we wouldn't want this for other
flox-activations subcommands