feat: add ensure_generation_locked with tests
Reimplements the functionality from `managed_environment`
and separates it from the lcokfile maangement.
In this implementation it does not read or write the lock but leaves
both to the caller.
This separates the floxmeta_branch implementation
from details of managed envrionments.

The tests are largely copied from the `managed_environment::tests`,
with a little bit polished documentation for their behaviour.