feat: add FloxmetaBranch::new() with integration tests
Tie together floxmeta creation, branch locking,
and setup of the "isntance branch".
This returns a validated branch instance and a lockfile.
The former should eventually implement all floxhub
git communcation behavior, and provide an interface to modify
the backing storage (i.e. `Generations`).
The latter is written to the .flox dir to reproduce
the same environemnt state
(and inheriuted from the current implementation).