refactor: remove direct uses of floxmeta in `managed_environment`
Avoid accessing floxmeta directly from managed environments.
The ultimate goal is to avoid raw git interactions
in managed environment entirely.
Push and Pull are currently unaffected to contain the size of the change.

Drpping the use of flooxmeta means `ManagedEnvironment`
does not need to carry both the branch and the repo,
nor does it need to dig as deep int floxmeta branch
to get a git instance.