feat: add an `--upstream` flag to `flox list`
When addressing managed envrionements (remote, or local) `--upstream`
can be provided to read the state of the environement on FloxHub.

Accessing remote data is still slightly awkward,
as we now have to destructure `ConcreteEnvironments` explicitly.
While i think this deserves some attention, (e.g. continuing the work on
separating floxmeta storage and reduciong envrionment type distinctions),
time sensitivity and locality to a few commands makes this less pressing
for the moment.