I considered a few different designs for this based on the following criteria: - A manifest that has `version = 1` and a package descriptor with an   `outputs` field should fail to parse. - Re-use as much existing functionality e.g. don't duplicate unaffected   data structures like `PackageDescriptorFlake`. - Don't change the lockfile format (yet), meaning that we shouldn't   change how package descriptors serialize. - Unambiguously parse a package descriptor that doesn't have an   `outputs` field.
I've given up on that last criterion because I don't think it's possible
without violating some of the other criteria.

The solution I came to was to rename the `Catalog` variant of
`ManifestPackageDescriptor` to `CatalogV1`, and then add a `CatalogV2`
variant that contains a `PackageDescriptorCatalogV2` struct. This has
the following implications:
- Any package descriptor that doesn't contain `outputs` will parse as a
  v1 package descriptor (we convert them to v2 in
  `RawManifest::to_typed`).
- Parsing a raw manifest will cause it to be interpreted as a v2
  manifest if it contains `version = 2` (not implemented yet).
- After it is parsed into a typed manifest, you must do a validation
  check to ensure that a v1 manifest doesn't contain any of the v2
  package descriptor variants.

Alternatives that I considered:
- Make `ManifestPackageDescriptor` an enum with `V1` and `V2` variants,
  where the `V1` variant contains what `ManifestPackageDescriptor`
  contains today.
    - I decided against that because you can't unambiguously parse a
      v1 or v2 descriptor when `outputs` isn't present.
- Make one of the package descriptor enums "tagged" so that they at
  least serialize unambiguously.
    - I decided against this because it changes the serialized format of
      the manifest in the lockfile.
- Use a const generic (the `1` on `Version<1>`) on `Install` so we can
  unambiguously parse.
    - I decided against this because it doesn't actually help us, as
      far as I can tell. You would want `Install<1>` to contain
      `PackageDescriptorCatalogV1` and `Install<2>` to contain
      `PackageDescriptorCatalogV2`, but you can't actually do that with
      const generics.

So, I don't love the solution I've come up with, but I think it has the
fewest downsides.

Parsing failures became worse after this change. Since we're using an
untagged enum representation of the different manifest versions, adding
an unknown field no longer failed with "unknown field 'foo', expected
...", and instead errored with effectively "because this unknown field
is here, I don't know which Manifest variant to parse to."

I refactored how we go from string -> `RawManifest` -> `Manifest` in
order to make the parsing failures nixer for users and to allow a
post-parsing step to verify that a v1 manifest doesn't contain v2
package descriptors. Previously we would convert string -> `RawManifest`
in `RawManifest::from_str`, which would then internally call
`RawManifest::to_typed` to ensure that there weren't any errors,
dropping the typed manifest. In all existing cases we would do this on
the way to calling `RawManifest::to_typed` *again* just to get a
`Manifest`. I've eliminated that internal `to_typed` so that we only
deserialize to `Manifest` once. This is a behavior change, but I don't
think it's a significant one. Previously, an arbitrary string that
successfully parsed with `RawManifest::from_str` was guaranteed to be a
valid manifest. That's no longer the case, which I think is actually
truer to what `RawManifest` is intended to be: valid TOML that loosely
resembles a manifest of some sort.

This refactor required updating some tests and rewiring some error
handling.