feat(activations): Change activations state dir
So that we have one `activations.json` and later one `executive` per
Flox environment, as defined by a `.flox` directory on disk. For remote
environments this is the local cache of the environment.

There's no need to perform a migration of the `Activations` schema as we
modify it in the next phase of the activation refactor for the
`executive` because older `flox` and `flox-watchdog` versions will
operate on the old state paths independently of new activations.

This requires us to temporarily disable the test for attaching to an
activation from a previous version of Flox because they are deliberately
disconnected. We can re-enable it after these changes are
released (v1.9.0?).

Whilst we're making a clean break we can also make some opportunistic
changes to how the path is constructed:

- nest under `{runtime_dir}/activations` so that they don't get mixed up
  with GC roots for remote environments in `{runtime_dir}/{owner}`

- suffix the hash with `-{basename(.flox)}` so that we have a human
  recognisable identifier, which is immutable because moving the dir
  results in a new `activations.json`, for debugging the contents of the
  directory