feat: allow invalidating cached remote state of a FloxHub environment
Since remote state is only updated asynchronously upon activation,
an activation following a push,
would incorrectly notify users of outstanding changes yet to be pushed,
as the cached remote state from previous activations is now outdated.
This function can be used to drop the cached remote state,
causing updates to be fetched again on a future activation
of the environment.

Note that this does not affect the remote state of the same environemnt
referenced in other places.
E.g. calling `invalidate_cached_remote_state(<env_in_dir_1)`
will not affect the data about and notifications notifications
for `<env_in_dir_2>`.