fix: don't cross filesystem boundaries renaming $out
The following Flox discourse issue highlights an instance where we
were attempting to move $out across filesystem boundaries, which of
course is implemented as a copy and subsequent removal within `mv`,
and the removal part of that operation could fail when the path
being removed contained directories with mode 555 (which is the
default for Nix).

This is fixed by ensuring that we use a temporary path on the same
filesystem as $out, which is most easily accomplished by simply
tagging a suffix to the path which moves it aside. This of course
relies on the ability to create new paths within `/nix/store`, which
I've confirmed is possible within the build sandbox on both Linux
and Darwin.

Also took this opportunity to precede _all_ builder invocations of
`rm -rf` with a prophylactic `find <path> -type d -exec chmod +w \;`.

See: https://discourse.flox.dev/t/pure-build-of-npm-package/1222