fix(path-hash): Use .flox
To match the new hashing scheme used when generating the parent
directory of `activations.json`.

Now detects the current environment or canonicalizes the provided path
to reduce the chance of accidentally providing the wrong path, as I've
done many times, which results in a hash that doesn't relate to anything
in use by code.