test: update manifest tests to use v2 manifests
I've updated existing tests and added a couple of new ones. The only
thing of note here is that we generate v1 manifests with only v1
package descriptors (no behavior change), and generate v2 manifests
with only v2 package descriptors. This is done because we convert all
v1 descriptors to v2 descriptors as part of parsing to the `Manifest`
type.

Following this may be confusing, so here's a couple of examples of how manifests will parse now:

```
version = 1

[install]
foo.pkg-path = "foo"
foo.outputs = "all"
```

This is strictly an error, the `outputs` field is not allowed in v1
manifests.

```
version = 2

[install]
foo.pkg-path = "foo"
foo.outputs = "all"
```

This unambiguously parses as a v2 package descriptor where `outputs` is
`Some(SelectedOutputs::All)`.

```
version = 2

[install]
foo.pkg-path = "foo"
```

This parses as:
- `Manifest::from_str`
    - calls `RawManifest::from_str`
    - calls `RawManifest::to_typed`
        - Converts v1 package descriptors to v2 package descriptors.
        - This is also where an error is thrown when a v1 manifest
          contains a v2 descriptor.