feat: alias `include.environments[].remote` as `floxhub-env`
Flox v1.8.0 replaced the concept of remote environments on the CLI level
with "local copies of FloxHub environments".
This leaves a consistency gap between the CLI and the manifest,
which still refers to "remote" [environments].

This commit adds a `floxhub-env` alias in its place,
which allows hand written manifests to be more descriptive.

When serialized into a lockfile `floxhub-env` will be written as `remote`
to ensure backwards compatibility of _lockfiles_ with Flox <= 1.8.0.
We aim to make breaking changes to the lockfile and manifest
in a future release which would rename the field
and implement migrations.