refactor: factor out AttachCtx
The executive will need to call start_services_blocking with information
that has not changed from when the executive was first started to when
the executive is signalled to start services. Since store path may have
changed since then, it should not be included.

There are a few other fields we don't need for start services that we
can leave off as well.

I'm not loving having ActivateCtx, ExecutiveCtx, and AttachCtx and
having those nested. There might be a better final layout, but I think
it's okay for now.