refactor: make manifest version dynamic
In order to make supporting "outputs to install" a smaller change, and
since it was decided that we won't persist "selected outputs" to the
locked packages in the lockfile, we need to support more than one
manifest version with a single data structure. That means removing the
const generic `Version` type and instead making the `version` field a
normal unisgned integer (u8 in this case since it's unlikely we'll have
more than 255 manifest versions).