build: replace `runCommandNoCC` alias with `runCommand`
The former is in fact an alias of `runCommand`.
Many of these aliases are being phased out
and dropped in a future release of nixpkgs.

See: <https://github.com/NixOS/nixpkgs/pull/456065>

Aliases are going to be replaced by a more automated/structural approach,
but that is orthogonal to the deprecation.